refactor(catalog): split patches.py into a layered package
The data-patch adapter had grown to 2325 lines in one module. Split it
into a package along the pipeline's dependency layers, behavior-preserving
(every definition body is byte-identical to the original):

- _types  — carriers shared across layers (PatchError, _Target, _CreatedKey)
- parsing — patch text -> PatchDoc (pure, no DB)
- emit    — low-level verbs that build plan rows for one resolved entry
- planning- build_plan: drives parsing output through emit, then validates

Dependencies flow one way: _types <- parsing <- emit <- planning. Each
phase-owned result type lives with its producer (e.g. _RemovalResult in
emit, _EntryResult in planning) rather than in a shared grab-bag. __init__
re-exports only the genuine public surface; the two helpers a test exercises
white-box are now imported from emit directly instead of via the package.
